+++ api/src/main/kotlin/org/kryptonmc/krypton/api/event/events/play/PermissionCheckEvent.kt
@@ -1,0 +1,33 @@
+package org.kryptonmc.krypton.api.event.events.play
+
+import org.kryptonmc.krypton.api.command.Sender
+import org.kryptonmc.krypton.api.event.Event
+
+data class PermissionCheckEvent(
+    val sender: Sender,
+    val permission: String,
+    val result: PermissionCheckResult
+) : Event
+
+/**
+ * The result of a permission check
+ *
+ * @author Callum Seabrook
+ */
+enum class PermissionCheckResult(val value: Boolean) {
+
+    /**
+     * Indicates this permission is set and true
+     */
+    TRUE(true),
+
+    /**
+     * Indicates this permission is set and false
+     */
+    FALSE(false),
+
+    /**
+     * Indicates this permission is unset
+     */
+    UNSET(false)
+}