+++ server/src/main/kotlin/org/kryptonmc/krypton/command/KryptonSender.kt
@@ -4,12 +4,12 @@ import org.kryptonmc.krypton.api.command.Sender
 
 abstract class KryptonSender : Sender {
 
-    private val permissions = mutableListOf<String>()
+    override val permissions = mutableMapOf<String, Boolean>()
 
-    override fun hasPermission(permission: String) = permission in permissions
+    override fun hasPermission(permission: String) = permission in permissions && permissions.getValue(permission)
 
     override fun grant(permission: String) {
-        permissions += permission
+        permissions[permission] = true
     }
 
     override fun grant(vararg permissions: String) = permissions.forEach { grant(it) }
@@ -17,7 +17,7 @@ abstract class KryptonSender : Sender {
     override fun grant(permissions: Iterable<String>) = permissions.forEach { grant(it) }
 
     override fun revoke(permission: String) {
-        permissions -= permission
+        permissions[permission] = false
     }
 
     override fun revoke(vararg permissions: String) = permissions.forEach { revoke(it) }